rename MailEditorN to MailEditor
not related to this issue, but it has bugged me ever since we deleted the original MailEditor and didnt rename this
