+++ src/contacts/ContactViewer.js
@@ -24,7 +24,7 @@ import {isoDateToBirthday} from "../api/common/utils/BirthdayUtils"
 import type {Contact} from "../api/entities/tutanota/Contact"
 import type {ContactSocialId} from "../api/entities/tutanota/ContactSocialId"
 import {locator} from "../api/main/MainLocator"
-import {newMailEditorFromTemplate} from "../mail/MailEditorN"
+import {newMailEditorFromTemplate} from "../mail/MailEditor"
 import {neverNull} from "../api/common/utils/Utils"
 import {logins} from "../api/main/LoginController"
