+++ src/mail/MailViewer.js
@@ -118,7 +118,7 @@ import {EventBanner} from "./EventBanner"
 import {checkApprovalStatus} from "../misc/LoginUtils"
 import {getEventFromFile} from "../calendar/CalendarInvites"
 import type {CalendarEvent} from "../api/entities/tutanota/CalendarEvent"
-import {newMailEditorAsResponse, newMailEditorFromDraft, newMailtoUrlMailEditor} from "./MailEditorN"
+import {newMailEditorAsResponse, newMailEditorFromDraft, newMailtoUrlMailEditor} from "./MailEditor"
 import type {MailboxDetail} from "./MailModel"
 import type {ResponseMailParameters} from "./SendMailModel"
 import {defaultSendMailModel} from "./SendMailModel"
