+++ src/subscription/giftcards/GiftCardUtils.js
@@ -32,7 +32,7 @@ import {getByAbbreviation} from "../../api/common/CountryList"
 import {NotAuthorizedError, NotFoundError} from "../../api/common/error/RestError"
 import {CancelledError} from "../../api/common/error/CancelledError"
 import {theme} from "../../gui/theme"
-import {writeGiftCardMail} from "../../mail/MailEditorN"
+import {writeGiftCardMail} from "../../mail/MailEditor"
 import {DefaultAnimationTime} from "../../gui/animation/Animations"
 import {copyToClipboard} from "../../misc/ClipboardUtils"
 import {BootIcons} from "../../gui/base/icons/BootIcons"
